+++ src/vue/forms/LoginForm.vue
@@ -27,6 +27,21 @@
         />
       </div>
     </div>
+    <div
+      v-if="tfaError"
+      class="app__form-row"
+    >
+      <div class="app__form-field">
+        <input-field
+          v-model="form.tfaCode"
+          @blur="touchField('form.tfaCode')"
+          id="login-tfa-code"
+          :error-message="getFieldErrorMessage('form.tfaCode')"
+          :white-autofill="false"
+          :label="'auth-pages.tfa-code' | globalize"
+        />
+      </div>
+    </div>
     <div class="app__form-actions">
       <button
         v-ripple
@@ -43,7 +58,7 @@
 <script>
 import FormMixin from '@/vue/mixins/form.mixin'
 
-import { required } from '@validators'
+import { required, requiredIf } from '@validators'
 import { vuexTypes } from '@/vuex'
 import { mapActions, mapGetters } from 'vuex'
 import { vueRoutes } from '@/vue-router/routes'
@@ -60,12 +75,17 @@ export default {
     form: {
       email: '',
       password: '',
+      tfaCode: '',
     },
+    tfaError: null,
   }),
   validations: {
     form: {
       email: { required },
       password: { required },
+      tfaCode: {
+        required: requiredIf(function () { return this.tfaError }),
+      },
     },
   },
   computed: {
@@ -84,6 +104,12 @@ export default {
       if (!this.isFormValid()) return
       this.disableForm()
       try {
+        if (this.tfaError) {
+          await Sdk.api.factors.verifyTotpFactor(
+            this.tfaError,
+            this.form.tfaCode
+          )
+        }
         await this.loadWallet({
           email: this.form.email.toLowerCase(),
           password: this.form.password,
@@ -119,10 +145,19 @@ export default {
             },
           })
           break
+        case errors.TFARequiredError:
+          this.tfaError = error
+          break
         case errors.NotFoundError:
           ErrorHandler.process(
             error,
             'auth-pages.wrong-email-or-password-err'
+          )
+          break
+        case errors.BadRequestError:
+          ErrorHandler.process(
+            error,
+            'auth-pages.wrong-tfa-code-err'
           )
           break
         default: